// src/core/input/input-enums.ts
export enum InputEventType {
    mousemove = "mousemove",
    mousedown = "mousedown",
    mouseup = "mouseup",
    mousewheel = "mousewheel",
    keyDown="keyDown",
    keyUp="keyUp",
}

export enum MouseInputType {
    leftButton = "leftButton",
    middleButton = "middleButton",
    rightButton = "rightButton",
    scrollUp = "scrollUp",
    scrollDown = "scrollDown"
}